Added optional parameter 'maxOccurrences' to replace functions
if 'maxOccurrences' is -1, which is the default, it replaces all occurrances. Can be optionally specified to only replace first 'maxOccurrences' ones. I couldn't think of a nice, and efficient way to implement this in multiReplace, to keep it being one iteration... I could implement it using a table to track how many times it replaced given strings, or I could just implement a 'multiReplacesFirsts' function