use `strerror_r` instead of `strerror`
`osErrorMsg` uses `strerror` on POSIX. However, `strerror` may not be
thread-safe. Therefore, the use of `strerror_r` is preferred. This patch
changes the implementation to use `strerror_r` instead.
Note that multiple signatures of `strerror_r` exist. The implementation
works for the POSIX definition (XSI-compatible) as well for the popular
`glibc` library when configured with `_GNU_SOURCE`.