Integrate OpenWatcom support into nim.cfg and nimbase.h
`nimbase.h` is modified in order to support the OpenWatcom C compiler
and it's pecularities, such as lack of support for `__attribute__`,
for getting/setting alignment, and thread-local variables, as well as
properly specifying the size of the various integer types.

Additionally in `nim.cfg`, `owcc`, OpenWatcom's POSIX compliant and
mostly GCC compatible compiler driver is specified as the default
cross compiler for the i086.dos and i086big.dos targets.