Define owcc (OpenWatcom C compiler) as a supported compiler.
This is a much better idea than simply changing the the `gcc` binary
to be `owcc`, since OpenWatcom does not support certain features such
as case ranges. `$nim/config/nim.cfg` has been updated to reflect this
change. The proper memory model is also now automatically selected for
the `i086` and `i086big` CPU targets and `-bdos` is also automatically
passed to the C compiler.