Add a new batteries module (which deprecates the prelude include)
Add a new "batteries" module which is similar to the existing prelude module which it deprecates. This was proposed as part of the discussion of
https://github.com/nim-lang/Nim/pull/20022

The main differences with prelude are that this module must be imported (while prelude had to be included), and that it imports a few more modules than prelude did.

This change also marks the prelude module as "deprecated". It does so by simply adding a "deprecation comment" to the prelude documentation. It also fixes an incorrect comment (which did not include strformat in the list of imported modules).
The reason why we cannot use the {.deprecated.} pragma is that prelude is included rather than imported, and thus the deprecated pragma would apply to the module that included prelude, rather than to the prelude module itself.