Avoid `static: test()` failure case by lifting block for just this new proc to global scope.
As mentioned in https://github.com/nim-lang/Nim/pull/21325 , this is a
more elemental VM failure, unrelated to `parseSize`.  As such, it needs
its own failure exhibition issue that is also a much smaller test case.
(I am working on that, but unless there is a burning need to `parseSize`
at compile time before run-time it need not hold up this PR.)