Warn when thread raises exceptions
When exceptions are not caught in thread procedures, they take down the
application (just like when it happens on the main thread).

However, when it happens on the main thread, unwinding still happens
meaning that destructors run and have a chance to perfrom cleanup as
usual - when it happens in a thread, all threads are abruptly
terminated.

This PR adds a compile-time warning when a thread is create that doesn't
catch its own exceptions. The warning can be turned off with
`-d:nimThreadsStrictRaises:off`