strict thread/tasks raises
Raising in threads causes the application to crash making it a
generally bad idea.

This PR takes the opportunity to restrict the errors raised by thread
proc's and tasks to only Defect, avoiding a common pitfall and source of
crashes.

The mode can be turned off for backwards compatiblity in a transition
period and a warning will be printed instead (useul for the version-1-6
version of this patch, so that code can start adapting).

Code that already doesn't raise won't be affected by this change - code
that potentially raises will see a compile error instead.

Regarding default on/off, a more compatible option is to enable thread
raise checking together with `nimPreviewSlimSystem` which already causes
compile errors for thread-using apps.

This PR disables strict error checking in a number of places and does
not attempt to fix the lack of error handling, ie nimgrep can still
crash - it merely highlights where such crashes will happen by using the
compatibility mode.

As can be seen in test suite changes, many functions are already
raises-free but need to be explicitly annotated as Nim is not smart
enough to figure this out in forward declarations and proc types.